move content builder into its own file
