+++ lib/src/builders/search/index.ts
@@ -1,6 +1,7 @@
 export * from "./facetBuilder";
 export * from "./searchBuilder";
 export * from "./searchRequestBuilder";
+export * from "./contentSearchBuilder";
 export * from "./productSearchBuilder";
 export * from "./searchCollectionBuilder";
 export * from "./searchTermPredictionBuilder";
